+++ includes/builtins.h
@@ -3,12 +3,6 @@
 
 # include "minishell.h"
 
-typedef struct s_cmd
-{
-	char	*path;
-	char	**cmd;
-}	t_cmd;
-
 void	mini_env(t_env *env);
 void	mini_pwd(void);
 void	mini_echo(t_cmd *cmd);
