+++ includes/init.h
@@ -3,18 +3,6 @@
 
 # include "minishell.h"
 
-typedef struct s_env_node
-{
-	char				*key;
-	char				*value;
-	struct s_env_node	*next;
-}	t_env_node;
-
-typedef struct s_env
-{
-	t_env_node h_node;
-} t_env;
-
 int	init_env(t_env *env, char **envp);
 
-# endif
+#endif
